Move splitAtSelection from util to device.
