+++ src/shortcuts/newThoughtOrOutdent.tsx
@@ -2,9 +2,10 @@ import React from 'react'
 import { Key } from 'ts-key-enum'
 import { Icon as IconType, Shortcut } from '../@types'
 import { isTouch } from '../browser'
-import { splitAtSelection, headValue, isDocumentEditable } from '../util'
+import { headValue, isDocumentEditable } from '../util'
 import { alert, newThought, outdent } from '../action-creators'
 import { isLastVisibleChild, simplifyPath } from '../selectors'
+import splitAtSelection from '../device/splitAtSelection'
 
 // eslint-disable-next-line jsdoc/require-jsdoc
 const Icon = ({ fill = 'black', size = 20, style }: IconType) => (
