+++ src/util/index.ts
@@ -89,7 +89,6 @@ export { setSelection } from './setSelection'
 export { sort } from './sort'
 export { spellNumber } from './spellNumber'
 export { splice } from './splice'
-export { splitAtSelection } from './splitAtSelection'
 export { splitSentence } from './splitSentence'
 export { storage } from './storage'
 export { strip } from './strip'
